gp_model/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
mediapipe>=0.8.0
numpy>=1.21.0
Pillow>=8.0.0
arabic-reshaper>=2.1.0
python-bidi>=0.4.2 